TEmPoSapi/requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.39.1
django-braces==1.13.0
django-filter==2.2.0
django-filters==0.2.1
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
idna==2.8
isort==4.3.20
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
ndg-httpsclient==0.5.1
oauthlib==3.0.1
Pillow==6.1.0
pyasn1==0.4.6
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pyOpenSSL==19.0.0
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
uritemplate==3.0.0
urllib3==1.25.2
wrapt==1.11.1
